legacy/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dexter"
version = "1.0.1"
description = "AI agent for financial analysis."
authors = []
requires-python = ">=3.10"
dependencies = [
    "googlenewsdecoder>=0.1.7",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.35",
    "langchain-anthropic>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langsmith>=0.3.37",
    "openai>=2.2.0",
    "prompt-toolkit>=3.0.0",
    "pydantic>=2.11.10",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]

[project.scripts]
dexter-agent = "dexter.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
